Added forum to devices page
